webapp/requirements_webapp.txt
python-dotenv>=1.0.0
llama-index-core>=0.10.0
llama-index-embeddings-huggingface>=0.1.0
llama-index-llms-openai>=0.1.0
openai>=1.12.0 
streamlit>=1.32.0